BDM_pro/requirements.txt
kafka-python
pymongo
streamlit
pandas
matplotlib